ENH: Use unsigned int for indexing into number of image dimensions
Replace non-portable "long" type with "int".

+++ Modules/Core/Common/include/itkNeighborhoodOperator.h
@@ -91,7 +91,7 @@ public:
 
   /** Sets the dimensional direction of a directional operator. */
   void
-  SetDirection(const unsigned long direction)
+  SetDirection(const unsigned int direction)
   {
     if (direction >= VDimension)
     {
@@ -102,7 +102,7 @@ public:
   }
 
   /** Returns the direction (dimension number) of a directional operator. */
-  unsigned long
+  unsigned int
   GetDirection() const
   {
     return m_Direction;
@@ -181,7 +181,7 @@ protected:
 
 private:
   /** Direction (dimension number) of the derivative. */
-  unsigned long m_Direction{ 0 };
+  unsigned int m_Direction{ 0 };
 };
 } // namespace itk
